+++ Changelog.md
@@ -13,6 +13,11 @@ _Also check out the [roadmap](Roadmap.md) with features I’d like to add in the
 -   New [Apple Events](https://rickies.co/apple-events) page with all Apple keynotes that were predicted in chronological order
 -   Improved navigation bar is now expandable to make room for the new pages
 
+### 10 March 2022
+
+-   Support for when other hosts match the donation of the Flexy loser, reflected on the [event page](https://rickies.co/keynote-mar-2022) and in the stats
+-   Support for the Flexy titles of the Duke, the Prince and the AG
+
 ### 13 January 2022
 
 -   The Annual Predictions of 2018 are differently styled on the timeline, because they were not official Rickies.